Fix error styles for each form control
If the control met the constraints, it will has green
boders, else it will has red borders and background color.

But, if it first met the constraints and then it's changed to not
meeting the constraints, it will still have the sucess style(green borders), which
is caused by forgetting to remove the (.success) class.

Instead of using .classList.add("success") and .classList("error"), replaced it
with .className = "success" and .className = "error" reduce the chance of forgeting
to add or remove classes.